// src/types/fieldtypes.ts
export enum FieldType {
	SINGLE_LINE_TEXT = 'singleLineText',
	EMAIL = 'email',
	URL = 'url',
	MULTILINE_TEXT = 'multilineText',
	NUMBER = 'number',
	PERCENT = 'percent',
	CURRENCY = 'currency',
	SINGLE_SELECT = 'singleSelect',
	MULTIPLE_SELECTS = 'multipleSelects',
	SINGLE_COLLABORATOR = 'singleCollaborator',
	MULTIPLE_COLLABORATORS = 'multipleCollaborators',
	MULTIPLE_RECORD_LINKS = 'multipleRecordLinks',
	DATE = 'date',
	DATE_TIME = 'dateTime',
	PHONE_NUMBER = 'phoneNumber',
	MULTIPLE_ATTACHMENTS = 'multipleAttachments',
	CHECKBOX = 'checkbox',
	FORMULA = 'formula',
	CREATED_TIME = 'createdTime',
	ROLLUP = 'rollup',
	COUNT = 'count',
	MULTIPLE_LOOKUP_VALUES = 'multipleLookupValues',
	AUTO_NUMBER = 'autoNumber',
	BARCODE = 'barcode',
	RATING = 'rating',
	RICH_TEXT = 'richText',
	DURATION = 'duration',
	LAST_MODIFIED_TIME = 'lastModifiedTime',
	CREATED_BY = 'createdBy',
	LAST_MODIFIED_BY = 'lastModifiedBy',
	BUTTON = 'button',
}